+++ packages/polymorphic/src/InferredPolymorphic/hooks/index.ts
@@ -1,0 +1,3 @@
+export { useInferredPolymorphic } from './useInferredPolymorphic';
+export { useInferredPolymorphicComponent } from './useInferredPolymorphicComponent';
+export { useInferredPolymorphicProps } from './useInferredPolymorphicProps';
