+++ packages/polymorphic/src/Polymorphic/constants.ts
@@ -1,0 +1,1 @@
+export const FALLBACK = 'div';
